Moved authentication to a separate class
The BeeHub_Auth class now handles authentication and all actions associated
with it. This keeps beehub_run.php clean and makes responsibilities more clear.
Also, when authenticating with an unknown SURFconext ID, BeeHub will now look
if we have a local account with the same e-mail address and if so, links the
SURFconext ID automatically to it.