Changed several properties to just store the variable part
Properties like default_sponsor (of a user) and owner (of a file) stored complete paths
to the principals. But as all principals of a certain type are always stored in the same
location and each property can only accept principals of one type, it is shorter to just
store the name instead of the full path.

Example:
DAV: owner = /system/home/niek
now becomes:
DAV: owner = niek
as an owner can only be a user (in BeeHub)