Default sponsor is now automatically set if needed
If the user is sponsored, but did not set his default sponsor yet, this is automatically done.
Also, when his default sponsor is not (no longer) one of his sponsors, the default sponsor is
changed. If the user doesn't have any sponsors at all (anymore), he also does never have a
default sponsor.