Reverted the (internal/shallow) locking system to the mySQL implementation
Because there were problems with the MongoDB implementation of the shallow
locks and my lack of time to create a propper solution, I choose to revert
back to the working state we had before. But only for the shallow locks, the
rest of the system will keep working with MongoDB. And extended filesystem
attributes are certainly not back in!